Add symbolizer option to generate debug information only.
There is no need to copy the .rodata and .text sections since the
consumers can get their data form the original ELF file. We mark
those section as SHT_NOBITS instead, which is the standard way to
create ELF file with separate debug information only.

The new option is called --only-keep-debug and it essentially has
the same effect as --only-keep-debug option for "strip" command.

Change-Id: I4f7488c04bbad86288beeda11c3bf295890e740d