Don't be so agressive in quering the code cache for new methods
Interim fix for apps which JIT a lot and trigger a lot of I/O because of
profile data (e.g. gmscore).

This reduces the amount of data re-written to disk by a few order of
magnitude. (e.g. 43k instead of 3.5MB)

Proper fix which will make the writing smarter is coming.

Bug:27600652
Change-Id: I7582bfab9abee6efebfe99da6979f7e118f82915