ART: Change prebuild run-test setup
Prebuild was meant to check whether the runtime correctly loads
pre-existing oat files. With the advent of patching images, this
has actually been reduced to "does the runtime run self-patchoat"
correctly.

We cannot produce a prebuilt oat file that does not have to be
patched, as the runtime will patch the image when it starts.

There are two options to make this more meaningful:

* In the prebuild setting, first run patchoat to explicitly create
  a valid dalvik-cache.

* Change the dex2oat output to be the odex file, which can be
  patched like a regular file.

This change shows the latter approach.

Change-Id: I5234e10d78f7ea6c7ad8598db67d3fad8ba42b67