ART:  Make mterp jit profiling race tolerant
The JIT profiling mechanism is intentionally non-precise to minimize
performance overhead.  In general, this is not a problem.   However,
the on-stack replacement mechanism assumes an order of method
compilation than can sometimes be violated if conditions are just
right.

This change allows compilation requests that were dropped due to
a race condition to eventually be re-issued.  It does this by allowing
the 16-bit hotness counter to wrap around.

Change-Id: I2ac8056af8c4f7f8cef3f2c3db70b0394c26a566