Revert "Revert "Fix ArtMethod::GetInvokeType for static methods on interfaces.""
bug:29089267
bug:27521545

We were hitting a compiler DCHECK that a class would never require to
do access checks on itself. The reason was that the compiler driver
was not trying to resolve a type, but instead relied on the verifier
for pre-populating the dex cache. However, the verifier doesn't
necessarily run in JIT mode.

This reverts commit 12abcbd950bd0ff4528e2e0d27ca5e881c7b0467.

(cherry picked from commit 3aaf9645639ff56fbfe49a981242cbda01d6796e)

Change-Id: I2f0de6e4f00f336dc61b2378054fb6d27a7298d0