Do not remove loads/store with unresolved accesses.
Due to AOT and compiling classes that are not ready yet
(eg missing a super class), we need to be conservative when
accessing fields and can end up in a situation where the same
field can be resolved and unresolved within the same method
(because of inlining). Therefore, disable removing loads and
stores in a method when there are unresolved accesses.

bug:29433999

(cherry picked from commit b93a16517dce12b83e9d5d3b599446e25fcc157a)

Change-Id: I8fcfb52c584222474a8220eb16c6581350b702e0