Fix System.arraycopy when doing same array copying.
At compile time, if constant source < constant destination, and we don't
know if the arrays are the same, then we must emit code that checks
if the two arrays are the same. If so, we jump to the slow path.

test:610-arraycopy

bug:30030084

(cherry picked from commit 9f65db89353c46f6b189656f7f55a99054e5cfce)

Change-Id:Ida67993d472b0ba4056d9c21c68f6e5239421f7d