[Xposed] Calculate called methods in compiler
This keeps track of all methods that are called from compiled code,
including inlined ones. Per compiled method, it stores the hashes of the
callees. In most cases, we could actually store the 16-bit method index
instead, however this would result in additional complexity for handling
callees which don't have a method entry in the dex file.