[Xposed] Add --xposed-only option to dex2oat
This option is useful create a file with just the Xposed information for
an .oat file. Especially for system apps and the boot image, we don't
want to store the whole recompiled .oat file, so we keep using the
original .oat file along with an extra file with just the Xposed
information.

The input .oat file is expected in the --dex-file parameter, the output
file is passed via the --oat-file parameter. This keeps the general
input/output semantics of the parameters, even though the names don't
really match the function anymore.