[Xposed] Function for invalidating callers of a method
Invalidate all callers of a method by checking the information from the
OatXposedFile. We need to check in all loaded dex files as there might
be cross-references. Also invalidate all JIT methods which call the
method. For callers which aren't loaded yet, remember the hash of the
callee and check for each newly loaded method whether it calls on of
those callees. If yes, set the method modifier bit for ignoring the AOT
code immediately.