[Xposed] Clone objects directly to the target class
Cloning an object with a different size causes some checks to fail. Even
though the proper target class will be set almost immediately, this
sitation can occur, especially under high load and with certain debug
features enabled. So better create the new object with the correct
class. Requires a change in the Xposed project as well.